Move deprecated components into deprecated folder + minor cleanup
