--- packages/core/src/components/Justification.ts
@@ -1,8 +1,0 @@
-export type justificationType =
-  | "flex-start"
-  | "flex-end"
-  | "center"
-  | "space-between"
-  | "space-around"
-  | "space-evenly"
-  | undefined;
