--- packages/core/src/components/ResizeMode.ts
@@ -1,7 +1,0 @@
-export type ResizeModeType =
-  | "cover"
-  | "contain"
-  | "stretch"
-  | "repeat"
-  | "center"
-  | undefined;
