+++ packages/core/src/deprecated-components/AnimatedCircularProgress.tsx
