+++ packages/core/src/deprecated-components/CircularProgress.tsx
