+++ packages/core/src/deprecated-components/DeprecatedFAB.tsx
@@ -2,7 +2,6 @@ import * as React from "react";
 import {
   ActivityIndicator,
   View,
-  Text,
   StyleSheet,
   StyleProp,
   ViewStyle,
@@ -11,41 +10,45 @@ import {
   PressableProps,
 } from "react-native";
 import color from "color";
-import Config from "./Config";
-import Elevation from "./Elevation";
+import Config from "../components/Config";
+import Text from "../components/Text";
+import Elevation from "../components/Elevation";
 import { withTheme } from "../theming";
 
 import type { Theme } from "../styles/DefaultTheme";
 import type { IconSlot } from "../interfaces/Icon";
 
 /**
- * A button is component that the user can press to trigger an action.
+ * A floating action button represents the primary action in an application.
  *
  * <div class="screenshots">
- *   <figure>
- *     <img src="screenshots/button-1.png" />
- *     <figcaption>Text button</figcaption>
- *   </figure>
- *   <figure>
- *     <img src="screenshots/button-2.png" />
- *     <figcaption>Outlined button</figcaption>
- *   </figure>
- *   <figure>
- *     <img src="screenshots/button-3.png" />
- *     <figcaption>Contained button</figcaption>
- *   </figure>
+ *   <img src="screenshots/fab-1.png" />
+ *   <img src="screenshots/fab-2.png" />
  * </div>
  *
  * ## Usage
  * ```js
  * import * as React from 'react';
- * import { Button } from '@draftbit/ui';
+ * import { StyleSheet } from 'react-native';
+ * import { FAB } from '@draftbit/ui';
  *
  * const MyComponent = () => (
- *   <Button icon="add-a-photo" type="solid" onPress={() => console.log('Pressed')}>
- *     Press me
- *   </Button>
+ *   <FAB
+ *     style={styles.fab}
+ *     type="outline"
+ *     icon="add"
+ *     onPress={() => console.log('Pressed')}
+ *   />
  * );
+ *
+ * const styles = StyleSheet.create({
+ *   fab: {
+ *     position: 'absolute',
+ *     margin: 16,
+ *     right: 0,
+ *     bottom: 0,
+ *   },
+ * })
  *
  * export default MyComponent;
  * ```
@@ -53,43 +56,45 @@ import type { IconSlot } from "../interfaces/Icon";
 
 type Props = {
   disabled?: boolean;
-  type?: "solid" | "outline" | "text";
+  type?: "solid" | "extended" | "outline" | "fixed" | "standard";
   loading?: boolean;
   icon?: string;
-  labelColor?: string;
   color?: string;
-  children?: React.ReactNode;
+  label?: string;
   onPress: () => void;
   elevation?: number;
-  style?: StyleProp<ViewStyle>;
   theme: Theme;
+  style?: StyleProp<ViewStyle>;
 } & PressableProps &
   IconSlot;
 
-const Button: React.FC<React.PropsWithChildren<Props>> = ({
+ /**
+ * @deprecated DEPRECATED
+ */
+const FAB: React.FC<React.PropsWithChildren<Props>> = ({
   Icon,
   icon,
   disabled = false,
   type = "solid",
   loading = false,
-  labelColor,
   color: colorOverride,
-  children,
+  label,
   onPress,
   elevation = 0,
+  style,
   theme: { colors, disabledOpacity, roundness, typography },
   ...rest
 }) => {
   let backgroundColor, borderColor, textColor, borderWidth;
   const buttonColor = colorOverride || colors.primary;
 
-  if (type === "solid") {
+  if (type === "standard" || type === "extended" || type === "fixed") {
     backgroundColor = buttonColor;
 
     if (disabled) {
       textColor = color(colors.surface).alpha(disabledOpacity).rgb().string();
     } else {
-      textColor = labelColor || colors.surface;
+      textColor = colors.surface;
     }
   } else {
     backgroundColor = "transparent";
@@ -97,7 +102,7 @@ const Button: React.FC<React.PropsWithChildren<Props>> = ({
     if (disabled) {
       textColor = color(buttonColor).alpha(disabledOpacity).rgb().string();
     } else {
-      textColor = labelColor || buttonColor;
+      textColor = buttonColor;
     }
   }
 
@@ -113,38 +118,67 @@ const Button: React.FC<React.PropsWithChildren<Props>> = ({
     borderWidth = 0;
   }
 
-  const buttonStyle = {
+  const buttonStyle: StyleProp<ViewStyle> = {
     backgroundColor,
     borderColor,
     borderWidth,
     borderRadius: roundness,
+    alignItems: "center",
+    justifyContent: "center",
   };
+
+  const buttonStyles: StyleProp<ViewStyle>[] = [styles.button, buttonStyle];
+
+  const contentStyle: StyleProp<ViewStyle>[] = [styles.content];
 
   const textStyle: StyleProp<TextStyle> = {
     textAlign: "center",
     color: textColor,
-    marginVertical: 16,
-    marginHorizontal: 16,
   };
 
-  const iconStyle = [
+  const iconStyle: StyleProp<ViewStyle>[] = [
     styles.icon,
     {
-      marginLeft: 16,
-      marginRight: -8,
       width: Config.buttonIconSize,
     },
   ];
 
+  if (type === "standard" || type === "outline") {
+    buttonStyle.width = Config.FABSize;
+    buttonStyle.height = Config.FABSize;
+    buttonStyle.borderRadius = Config.FABBorderRadius;
+
+    contentStyle.push({
+      width: Config.FABSize,
+      height: Config.FABSize,
+    });
+  }
+
+  if (type === "extended" || type === "fixed") {
+    iconStyle.push({
+      marginLeft: 16,
+      marginRight: -8,
+    });
+
+    textStyle.margin = 16;
+  }
+
+  if (type === "fixed") {
+    buttonStyles.push({
+      height: Config.FABFixedHeight,
+      alignSelf: "stretch",
+    });
+  }
+
   return (
-    <Elevation style={{ elevation, alignSelf: "stretch" }}>
+    <Elevation style={[{ elevation }, style]}>
       <Pressable
         {...rest}
         onPress={onPress}
         accessibilityState={{ disabled }}
         accessibilityRole="button"
         disabled={disabled || loading}
-        style={[styles.button, buttonStyle]}
+        style={buttonStyles}
       >
         <View style={styles.content}>
           {icon && loading !== true ? (
@@ -163,9 +197,11 @@ const Button: React.FC<React.PropsWithChildren<Props>> = ({
               style={iconStyle}
             />
           ) : null}
-          <Text numberOfLines={1} style={[textStyle, typography.button]}>
-            {children}
-          </Text>
+          {label ? (
+            <Text numberOfLines={1} style={[textStyle, typography.button]}>
+              {label}
+            </Text>
+          ) : null}
         </View>
       </Pressable>
     </Elevation>
@@ -174,7 +210,6 @@ const Button: React.FC<React.PropsWithChildren<Props>> = ({
 
 const styles = StyleSheet.create({
   button: {
-    minWidth: 64,
     borderStyle: "solid",
   },
   content: {
@@ -183,8 +218,17 @@ const styles = StyleSheet.create({
     justifyContent: "center",
   },
   icon: {
+    alignItems: "center",
+    justifyContent: "center",
     width: Config.buttonIconSize,
+  },
+  fixed: {
+    left: 0,
+    right: 0,
+    bottom: 0,
+    height: 64,
+    borderRadius: 0,
   },
 });
 
-export default withTheme(Button);
+export default withTheme(FAB);
